--- a/const.go
+++ b/const.go
@@ -14,13 +14,11 @@
 
 package apd
 
-import "math/big"
-
 var (
-	bigOne  = big.NewInt(1)
-	bigTwo  = big.NewInt(2)
-	bigFive = big.NewInt(5)
-	bigTen  = big.NewInt(10)
+	bigOne  = NewBigInt(1)
+	bigTwo  = NewBigInt(2)
+	bigFive = NewBigInt(5)
+	bigTen  = NewBigInt(10)
 
 	decimalZero      = New(0, 0)
 	decimalOneEighth = New(125, -3)