--- a/decimal.go
+++ b/decimal.go
@@ -15,12 +15,11 @@
 package apd
 
 import (
-	"database/sql/driver"
 	"math"
-	"math/big"
 	"strconv"
 	"strings"
 
+	"database/sql/driver"
 	"github.com/pkg/errors"
 )
 
@@ -34,7 +33,7 @@ type Decimal struct {
 	Form     Form
 	Negative bool
 	Exponent int32
-	Coeff    big.Int
+	Coeff    BigInt
 }
 
 // Form specifies the form of a Decimal.
@@ -78,25 +77,20 @@ const (
 
 // New creates a new decimal with the given coefficient and exponent.
 func New(coeff int64, exponent int32) *Decimal {
-	d := &Decimal{
-		Negative: coeff < 0,
-		Coeff:    *big.NewInt(coeff),
-		Exponent: exponent,
-	}
-	d.Coeff.Abs(&d.Coeff)
+	d := new(Decimal)
+	d.SetFinite(coeff, exponent)
 	return d
 }
 
 // NewWithBigInt creates a new decimal with the given coefficient and exponent.
-func NewWithBigInt(coeff *big.Int, exponent int32) *Decimal {
-	d := &Decimal{
-		Exponent: exponent,
-	}
+func NewWithBigInt(coeff *BigInt, exponent int32) *Decimal {
+	d := new(Decimal)
 	d.Coeff.Set(coeff)
 	if d.Coeff.Sign() < 0 {
 		d.Negative = true
 		d.Coeff.Abs(&d.Coeff)
 	}
+	d.Exponent = exponent
 	return d
 }
 
@@ -243,7 +237,8 @@ func (d *Decimal) SetFloat64(f float64) (*Decimal, error) {
 	return d, err
 }
 
-// Int64 returns the int64 representation of x. If x cannot be represented in an int64, an error is returned.
+// Int64 returns the int64 representation of x. If x cannot be represented in an
+// int64, an error is returned.
 func (d *Decimal) Int64() (int64, error) {
 	if d.Form != Finite {
 		return 0, errors.Errorf("%s is not finite", d.String())
@@ -339,7 +334,7 @@ func (d *Decimal) setExponent(c *Context, res Condition, xs ...int64) Condition 
 				res |= Clamped
 			}
 			r = Etiny
-			d.Coeff = integ.Coeff
+			d.Coeff.Set(&integ.Coeff)
 			res |= Rounded
 		}
 	} else if v > c.MaxExponent {
@@ -360,10 +355,10 @@ func (d *Decimal) setExponent(c *Context, res Condition, xs ...int64) Condition 
 	return res
 }
 
-// upscale converts a and b to big.Ints with the same scaling. It returns
+// upscale converts a and b to BigInts with the same scaling. It returns
 // them with this scaling, along with the scaling. An error can be produced
 // if the resulting scale factor is out of range.
-func upscale(a, b *Decimal) (*big.Int, *big.Int, int32, error) {
+func upscale(a, b *Decimal) (*BigInt, *BigInt, int32, error) {
 	if a.Exponent == b.Exponent {
 		return &a.Coeff, &b.Coeff, a.Exponent, nil
 	}
@@ -378,7 +373,7 @@ func upscale(a, b *Decimal) (*big.Int, *big.Int, int32, error) {
 	if s > MaxExponent {
 		return nil, nil, 0, errors.New(errExponentOutOfRangeStr)
 	}
-	x := new(big.Int)
+	x := new(BigInt)
 	e := tableExp10(s, x)
 	x.Mul(&a.Coeff, e)
 	y := &b.Coeff
@@ -389,7 +384,7 @@ func upscale(a, b *Decimal) (*big.Int, *big.Int, int32, error) {
 }
 
 // setBig sets b to d's coefficient with negative.
-func (d *Decimal) setBig(b *big.Int) *big.Int {
+func (d *Decimal) setBig(b *BigInt) *BigInt {
 	b.Set(&d.Coeff)
 	if d.Negative {
 		b.Neg(b)
@@ -552,7 +547,7 @@ func (d *Decimal) Cmp(x *Decimal) int {
 		return gt
 	}
 
-	// Now have to use aligned big.Ints. This function previously used upscale to
+	// Now have to use aligned BigInts. This function previously used upscale to
 	// align in all cases, but that requires an error in the return value. upscale
 	// does that so that it can fail if it needs to take the Exp of too-large a
 	// number, which is very slow. The only way for that to happen here is for d
@@ -562,12 +557,12 @@ func (d *Decimal) Cmp(x *Decimal) int {
 
 	var cmp int
 	if d.Exponent < x.Exponent {
-		var xScaled big.Int
+		var xScaled BigInt
 		xScaled.Set(&x.Coeff)
 		xScaled.Mul(&xScaled, tableExp10(int64(x.Exponent)-int64(d.Exponent), nil))
 		cmp = d.Coeff.Cmp(&xScaled)
 	} else {
-		var dScaled big.Int
+		var dScaled BigInt
 		dScaled.Set(&d.Coeff)
 		dScaled.Mul(&dScaled, tableExp10(int64(d.Exponent)-int64(x.Exponent), nil))
 		cmp = dScaled.Cmp(&x.Coeff)
@@ -644,7 +639,7 @@ func (d *Decimal) Modf(integ, frac *Decimal) {
 
 	e := tableExp10(exp, nil)
 
-	var icoeff *big.Int
+	var icoeff *BigInt
 	if integ != nil {
 		icoeff = &integ.Coeff
 		integ.Exponent = 0
@@ -652,7 +647,7 @@ func (d *Decimal) Modf(integ, frac *Decimal) {
 	} else {
 		// This is the integ == nil branch, and we already checked if both integ and
 		// frac were nil above, so frac can never be nil in this branch.
-		icoeff = new(big.Int)
+		icoeff = new(BigInt)
 	}
 
 	if frac != nil {
@@ -724,8 +719,8 @@ func (d *Decimal) Reduce(x *Decimal) (*Decimal, int) {
 
 	// Divide by 10 in a loop. In benchmarks of reduce0.decTest, this is 20%
 	// faster than converting to a string and trimming the 0s from the end.
-	z := d.setBig(new(big.Int))
-	r := new(big.Int)
+	z := d.setBig(new(BigInt))
+	r := new(BigInt)
 	for {
 		z.QuoRem(&d.Coeff, bigTen, r)
 		if r.Sign() == 0 {