--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
-module github.com/cockroachdb/apd/v2
+module github.com/cockroachdb/apd/v3
+
+go 1.13
 
 require github.com/pkg/errors v0.8.0